Extract non-empty-string substr NodeScopeResolver test cases and adapt them according to PHP version
